requirements.txt
discord.py
GitPython
